Move frange and percentageOf to slider_utils
This finally bothered me enough to fix
